3.0.0b1
This commit represents an effort made by the Checkout team to build a brand new SDK for Python. It ships a
new instantiation layer that makes easier a less confusing to set up the SDK. From the point of view of
usability, the new version follows similar principles, although major interfaces are now split/duplicated
to reference different data schemes which makes it fully compatible with `default` (MBC) and `Four` (NAS)
account structures. This separation is statically forced during instantiation.

The SDK supports the following modules:
* Tokens (default/four)
* Instruments (default/four)
* Customers (default/four)
* Disputes (default/four)
* Payments (default/four)
* Alternative Payments (default/four)
* Sources (default)
* Forex (four)
* Marketplace (four)
* Sessions (four)
* Hosted & Links Payments

`README.md` was updated to match the new setup. Code of conduct and license were reviewed.